+++ gems/quilt_rails/lib/quilt_rails/react_renderable.rb
@@ -30,19 +30,19 @@ module Quilt
 
     def proxy(headers, data)
       url = "#{Quilt.configuration.react_server_protocol}://#{Quilt.configuration.react_server_host}"
-      Quilt::Logger.log("[ReactRenderable] proxying to React server at #{url}")
+      Quilt.logger.info("[ReactRenderable] proxying to React server at #{url}")
 
       unless headers.blank?
-        Quilt::Logger.log("[ReactRenderable] applying custom headers #{headers.inspect}")
+        Quilt.logger.info("[ReactRenderable] applying custom headers #{headers.inspect}")
       end
 
       begin
         reverse_proxy(
           url,
-          headers: headers.merge('X-Quilt-Data': data.to_json)
+          headers: headers.merge('X-Request-ID': request.request_id, 'X-Quilt-Data': data.to_json)
         ) do |callbacks|
           callbacks.on_response do |status_code, _response|
-            Quilt::Logger.log("[ReactRenderable] #{url} returned #{status_code}")
+            Quilt.logger.info("[ReactRenderable] #{url} returned #{status_code}")
           end
         end
       rescue Errno::ECONNREFUSED
@@ -52,17 +52,19 @@ module Quilt
 
     class ReactServerNoResponseError < StandardError
       def initialize(url)
-        # rubocop:disable LineLength
-        super "Errno::ECONNREFUSED: Waiting for React server to boot up. If this error persists verify that @shopify/react-server is configured on #{url}"
-        # rubocop:enable LineLength
+        super(<<~MSG.squish)
+          Errno::ECONNREFUSED: Waiting for React server to boot up.
+          If this error persists verify that @shopify/react-server is configured on #{url}"
+        MSG
       end
     end
 
     class DoNotIntegrationTestError < StandardError
       def initialize
-        # rubocop:disable LineLength
-        super "Do not try to use Rails integration tests on your quilt_rails app. Instead use Jest and @shopify/react-testing to test your React application directly."
-        # rubocop:enable LineLength
+        super(<<~MSG.squish)
+          Do not try to use Rails integration tests on your quilt_rails app.
+          Instead use Jest and @shopify/react-testing to test your React application directly."
+        MSG
       end
     end
   end