+++ packages/address-mocks/src/fixtures/country_ca_af.ts
@@ -19,6 +19,9 @@ const data = {
         postalCode: 'Postal code',
         zone: 'Province',
       },
+      optionalLabels: {
+        address2: 'Apt./Unit No. (optional)',
+      },
       formatting: {
         edit:
           '{firstName}{lastName}_{company}_{address1}_{address2}_{city}_{country}{province}{zip}_{phone}',