+++ packages/koa-shopify-auth/src/auth/redirection-page.ts
@@ -4,18 +4,18 @@ export default function redirectionScript({origin, redirectTo, apiKey}) {
       document.addEventListener('DOMContentLoaded', function() {
         if (window.top === window.self) {
           // If the current window is the 'parent', change the URL by setting location.href
-          window.location.href = '${redirectTo}';
+          window.location.href = "${redirectTo}";
         } else {
           // If the current window is the 'child', change the parent's URL with postMessage
           var AppBridge = window['app-bridge'];
           var createApp = AppBridge.default;
           var Redirect = AppBridge.actions.Redirect;
           var app = createApp({
-            apiKey: '${apiKey}',
-            shopOrigin: '${origin}',
+            apiKey: "${apiKey}",
+            shopOrigin: "${encodeURI(origin)}",
           });
           var redirect = Redirect.create(app);
-          redirect.dispatch(Redirect.Action.REMOTE, '${redirectTo}');
+          redirect.dispatch(Redirect.Action.REMOTE, "${redirectTo}");
         }
       });
     </script>