+++ packages/address-mocks/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [1.4.0] - 2020-09-02
+
+- Updated mocks for countries and country calls to include `optionalLabels`. [#1610](https://github.com/Shopify/quilt/pull/1610)
 
 ## [1.3.0] - 2020-03-10