+++ packages/koa-shopify-webhooks/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [2.5.0] - 2020-08-20
+
+- Add an option to register EventBridge webhooks to `registerWebhook`.
+
 ## [2.4.0] - 2020-02-19
 
 - The `ApiVersion` enum now has an `January20` and `April20` options